feat: Task 4完了 - MobilePerformanceOptimizer.jsの分割
MobilePerformanceOptimizer.jsを8,405語から1,858語に大幅削減（78%削減）：

## 新規作成したコンポーネント
- **MobileResourceManager.js** - メモリリソース管理、テクスチャプール、アセットストリーミング
- **MobileRenderOptimizer.js** - レンダリング品質制御、フレームレート最適化
- **MobileBatteryOptimizer.js** - バッテリー監視、電力管理モード
- **MobileMemoryManager.js** - メモリ圧迫検知、GC管理、リーク検出

## 主要改善点
- 単一責任の原則に基づくコンポーネント分離
- 各コンポーネント間の依存関係注入パターン
- 統合アーキテクチャによる協調動作
- レガシーAPI互換性の維持
