+++ .kiro/specs/mcp-token-limit-fix/tasks.md
@@ -45,13 +45,14 @@
 
 ## 高優先度対応フェーズ
 
-- [ ] 4. MobilePerformanceOptimizer.js の分割（8,405語 → 複数ファイル）
+- [x] 4. MobilePerformanceOptimizer.js の分割（8,405語 → 1,858語）
   - MobilePerformanceOptimizer.js のメインオプティマイザー部分の抽出（< 2,500語）
   - MobileResourceManager.js の作成（リソース管理）
   - MobileRenderOptimizer.js の作成（レンダリング最適化）
   - MobileBatteryOptimizer.js の作成（バッテリー最適化）
   - MobileMemoryManager.js の作成（メモリ管理）
   - インポート文の更新と依存関係の修正
+  - 新コンポーネントアーキテクチャによる統合
   - 分割後のパフォーマンステスト実行
   - _要件: 1.3, 2.2, 2.3_
