+++ src/utils/mobile/MobileBatteryOptimizer.js
@@ -1,0 +1,629 @@
+/**
+ * MobileBatteryOptimizer.js
+ * モバイルバッテリー最適化システム
+ * MobilePerformanceOptimizerから分離されたバッテリー最適化機能
+ */
+
+import { getErrorHandler } from '../ErrorHandler.js';
+import { getConfigurationManager } from '../../core/ConfigurationManager.js';
+
+export class MobileBatteryOptimizer {
+    constructor() {
+        this.errorHandler = getErrorHandler();
+        this.configManager = getConfigurationManager();
+        
+        // Battery optimization configuration
+        this.batteryConfig = {
+            enabled: true,
+            
+            // Battery monitoring
+            monitoring: {
+                enabled: true,
+                updateInterval: 30000, // 30 seconds
+                lowBatteryThreshold: 0.20, // 20%
+                criticalBatteryThreshold: 0.10, // 10%
+                chargingDetection: true
+            },
+            
+            // Power management modes
+            powerModes: {
+                normal: {
+                    name: 'normal',
+                    frameRateLimit: 60,
+                    renderQuality: 'high',
+                    backgroundThrottling: false,
+                    networkOptimization: false,
+                    screenBrightness: 1.0
+                },
+                powersaver: {
+                    name: 'powersaver',
+                    frameRateLimit: 30,
+                    renderQuality: 'medium',
+                    backgroundThrottling: true,
+                    networkOptimization: true,
+                    screenBrightness: 0.7
+                },
+                extreme: {
+                    name: 'extreme',
+                    frameRateLimit: 15,
+                    renderQuality: 'low',
+                    backgroundThrottling: true,
+                    networkOptimization: true,
+                    screenBrightness: 0.5
+                }
+            },
+            
+            currentMode: 'normal',
+            autoModeSwitch: true,
+            
+            // Battery optimizations
+            optimizations: {
+                reducedFrameRate: false,
+                backgroundThrottling: true,
+                cpuThrottling: false,
+                screenDimming: false,
+                networkOptimization: true,
+                cacheAggressive: true,
+                suspendInactive: true,
+                minimizeWakeups: true
+            },
+            
+            // Thermal management for battery
+            thermalManagement: {
+                enabled: true,
+                temperatureThreshold: 40, // Celsius
+                cooldownRequired: false,
+                throttleOnHeat: true
+            }
+        };
+        
+        // Battery monitoring state
+        this.batteryMonitoring = {
+            enabled: false,
+            
+            // Battery API data
+            battery: {
+                level: 1.0,
+                charging: false,
+                chargingTime: Infinity,
+                dischargingTime: Infinity,
+                lastUpdate: Date.now()
+            },
+            
+            // Battery usage estimation
+            usage: {
+                currentDrain: 0, // mAh per hour
+                averageDrain: 0,
+                estimatedRemaining: 0, // hours
+                drainHistory: [],
+                maxHistorySize: 100
+            },
+            
+            // Power consumption tracking
+            powerConsumption: {
+                cpu: 0,
+                gpu: 0,
+                screen: 0,
+                network: 0,
+                total: 0,
+                baselinePower: 50 // mAh baseline consumption
+            },
+            
+            // Battery health estimation
+            health: {
+                capacity: 100, // percentage of original capacity
+                cycleCount: 0,
+                temperature: 25, // Celsius
+                voltage: 3.7 // Volts
+            }
+        };
+        
+        // Performance tracking for battery optimization
+        this.performanceTracking = {
+            frameRateHistory: [],
+            cpuUsageHistory: [],
+            gpuUsageHistory: [],
+            networkUsageHistory: [],
+            lastMeasurement: Date.now()
+        };
+        
+        // Initialize battery optimizer
+        this.initializeBatteryOptimizer();
+    }
+    
+    /**
+     * Initialize battery optimization system
+     */
+    async initializeBatteryOptimizer() {
+        console.log('[MobileBatteryOptimizer] Initializing battery optimization...');
+        
+        try {
+            await this.setupBatteryAPI();
+            this.setupPerformanceTracking();
+            this.startBatteryMonitoring();
+            this.applyInitialOptimizations();
+            
+            console.log('[MobileBatteryOptimizer] Battery optimization initialized successfully');
+        } catch (error) {
+            this.errorHandler.handleError(error, 'MobileBatteryOptimizer.initializeBatteryOptimizer');
+        }
+    }
+    
+    /**
+     * Setup Battery API if available
+     */
+    async setupBatteryAPI() {
+        try {
+            if ('getBattery' in navigator) {
+                const battery = await navigator.getBattery();
+                this.setupBatteryEventListeners(battery);
+                this.updateBatteryInfo(battery);
+                this.batteryMonitoring.enabled = true;
+                
+                console.log('[MobileBatteryOptimizer] Battery API initialized');
+            } else {
+                console.log('[MobileBatteryOptimizer] Battery API not available, using estimation');
+                this.setupBatteryEstimation();
+            }
+        } catch (error) {
+            console.warn('[MobileBatteryOptimizer] Failed to setup Battery API:', error);
+            this.setupBatteryEstimation();
+        }
+    }
+    
+    /**
+     * Setup battery event listeners
+     */
+    setupBatteryEventListeners(battery) {
+        battery.addEventListener('chargingchange', () => {
+            this.updateBatteryInfo(battery);
+            this.handleChargingStateChange(battery.charging);
+        });
+        
+        battery.addEventListener('levelchange', () => {
+            this.updateBatteryInfo(battery);
+            this.handleBatteryLevelChange(battery.level);
+        });
+        
+        battery.addEventListener('chargingtimechange', () => {
+            this.updateBatteryInfo(battery);
+        });
+        
+        battery.addEventListener('dischargingtimechange', () => {
+            this.updateBatteryInfo(battery);
+        });
+    }
+    
+    /**
+     * Update battery information
+     */
+    updateBatteryInfo(battery) {
+        const batteryData = this.batteryMonitoring.battery;
+        
+        batteryData.level = battery.level;
+        batteryData.charging = battery.charging;
+        batteryData.chargingTime = battery.chargingTime;
+        batteryData.dischargingTime = battery.dischargingTime;
+        batteryData.lastUpdate = Date.now();
+        
+        // Update power consumption estimation
+        this.updatePowerConsumption();
+    }
+    
+    /**
+     * Setup battery estimation for devices without Battery API
+     */
+    setupBatteryEstimation() {
+        // Use performance and usage patterns to estimate battery drain
+        this.batteryMonitoring.battery.level = 0.8; // Assume 80% initially
+        this.batteryMonitoring.battery.charging = false;
+        
+        console.log('[MobileBatteryOptimizer] Battery estimation mode enabled');
+    }
+    
+    /**
+     * Handle charging state change
+     */
+    handleChargingStateChange(charging) {
+        console.log(`[MobileBatteryOptimizer] Charging state changed: ${charging}`);
+        
+        if (charging) {
+            // Switch to normal mode when charging
+            this.setPowerMode('normal');
+        } else {
+            // Evaluate appropriate power mode when unplugged
+            this.evaluateOptimalPowerMode();
+        }
+    }
+    
+    /**
+     * Handle battery level change
+     */
+    handleBatteryLevelChange(level) {
+        console.log(`[MobileBatteryOptimizer] Battery level changed: ${(level * 100).toFixed(1)}%`);
+        
+        const config = this.batteryConfig.monitoring;
+        
+        if (level <= config.criticalBatteryThreshold) {
+            this.handleCriticalBattery();
+        } else if (level <= config.lowBatteryThreshold) {
+            this.handleLowBattery();
+        }
+        
+        // Auto-adjust power mode based on battery level
+        if (this.batteryConfig.autoModeSwitch) {
+            this.evaluateOptimalPowerMode();
+        }
+    }
+    
+    /**
+     * Handle critical battery level
+     */
+    handleCriticalBattery() {
+        console.warn('[MobileBatteryOptimizer] Critical battery level detected');
+        
+        // Force extreme power saving mode
+        this.setPowerMode('extreme');
+        
+        // Enable all battery optimizations
+        this.enableAllOptimizations();
+        
+        // Notify application
+        this.notifyBatteryState('critical');
+    }
+    
+    /**
+     * Handle low battery level
+     */
+    handleLowBattery() {
+        console.warn('[MobileBatteryOptimizer] Low battery level detected');
+        
+        // Switch to power saver mode
+        this.setPowerMode('powersaver');
+        
+        // Notify application
+        this.notifyBatteryState('low');
+    }
+    
+    /**
+     * Setup performance tracking for battery optimization
+     */
+    setupPerformanceTracking() {
+        setInterval(() => {
+            this.updatePerformanceMetrics();
+        }, 5000); // Update every 5 seconds
+        
+        console.log('[MobileBatteryOptimizer] Performance tracking initialized');
+    }
+    
+    /**
+     * Start battery monitoring
+     */
+    startBatteryMonitoring() {
+        if (!this.batteryConfig.monitoring.enabled) return;
+        
+        setInterval(() => {
+            this.updateBatteryUsage();
+            this.evaluateBatteryHealth();
+            this.optimizePowerConsumption();
+        }, this.batteryConfig.monitoring.updateInterval);
+        
+        console.log('[MobileBatteryOptimizer] Battery monitoring started');
+    }
+    
+    /**
+     * Apply initial battery optimizations
+     */
+    applyInitialOptimizations() {
+        // Start with normal mode
+        this.setPowerMode('normal');
+        
+        // Enable basic optimizations
+        this.applyBasicOptimizations();
+        
+        console.log('[MobileBatteryOptimizer] Initial optimizations applied');
+    }
+    
+    /**
+     * Set power management mode
+     */
+    setPowerMode(mode) {
+        if (!this.batteryConfig.powerModes[mode]) {
+            console.warn(`[MobileBatteryOptimizer] Invalid power mode: ${mode}`);
+            return;
+        }
+        
+        const previousMode = this.batteryConfig.currentMode;
+        const modeSettings = this.batteryConfig.powerModes[mode];
+        
+        this.batteryConfig.currentMode = mode;
+        this.applyPowerModeSettings(modeSettings);
+        
+        console.log(`[MobileBatteryOptimizer] Power mode changed: ${previousMode} → ${mode}`);
+    }
+    
+    /**
+     * Apply power mode settings
+     */
+    applyPowerModeSettings(settings) {
+        const opts = this.batteryConfig.optimizations;
+        
+        // Apply frame rate limit
+        opts.reducedFrameRate = settings.frameRateLimit < 60;
+        
+        // Apply background throttling
+        opts.backgroundThrottling = settings.backgroundThrottling;
+        
+        // Apply network optimization
+        opts.networkOptimization = settings.networkOptimization;
+        
+        // Apply screen brightness (if supported)
+        this.adjustScreenBrightness(settings.screenBrightness);
+        
+        console.log(`[MobileBatteryOptimizer] Power mode settings applied - Frame Rate: ${settings.frameRateLimit}, Quality: ${settings.renderQuality}`);
+    }
+    
+    /**
+     * Adjust screen brightness (if supported)
+     */
+    adjustScreenBrightness(brightness) {
+        // This would typically require native integration
+        // For web, we can suggest to the user or adjust CSS filters
+        
+        if (brightness < 1.0) {
+            console.log(`[MobileBatteryOptimizer] Screen brightness optimization suggested: ${(brightness * 100).toFixed(0)}%`);
+        }
+    }
+    
+    /**
+     * Evaluate optimal power mode based on current conditions
+     */
+    evaluateOptimalPowerMode() {
+        const battery = this.batteryMonitoring.battery;
+        const config = this.batteryConfig.monitoring;
+        
+        if (battery.charging) {
+            this.setPowerMode('normal');
+            return;
+        }
+        
+        if (battery.level <= config.criticalBatteryThreshold) {
+            this.setPowerMode('extreme');
+        } else if (battery.level <= config.lowBatteryThreshold) {
+            this.setPowerMode('powersaver');
+        } else if (battery.level > 0.5) {
+            this.setPowerMode('normal');
+        }
+    }
+    
+    /**
+     * Update performance metrics for battery optimization
+     */
+    updatePerformanceMetrics() {
+        const tracking = this.performanceTracking;
+        const now = Date.now();
+        
+        // Simulate performance metrics (would be actual values in real implementation)
+        const frameRate = 60 - Math.random() * 10; // 50-60 FPS
+        const cpuUsage = Math.random() * 0.8; // 0-80%
+        const gpuUsage = Math.random() * 0.6; // 0-60%
+        const networkUsage = Math.random() * 0.3; // 0-30%
+        
+        // Add to history
+        tracking.frameRateHistory.push({ time: now, value: frameRate });
+        tracking.cpuUsageHistory.push({ time: now, value: cpuUsage });
+        tracking.gpuUsageHistory.push({ time: now, value: gpuUsage });
+        tracking.networkUsageHistory.push({ time: now, value: networkUsage });
+        
+        // Limit history size
+        const maxHistorySize = 100;
+        if (tracking.frameRateHistory.length > maxHistorySize) {
+            tracking.frameRateHistory.shift();
+            tracking.cpuUsageHistory.shift();
+            tracking.gpuUsageHistory.shift();
+            tracking.networkUsageHistory.shift();
+        }
+        
+        tracking.lastMeasurement = now;
+        
+        // Update power consumption based on performance
+        this.updatePowerConsumption();
+    }
+    
+    /**
+     * Update power consumption estimation
+     */
+    updatePowerConsumption() {
+        const consumption = this.batteryMonitoring.powerConsumption;
+        const tracking = this.performanceTracking;
+        
+        if (tracking.cpuUsageHistory.length === 0) return;
+        
+        // Get latest metrics
+        const latestCpu = tracking.cpuUsageHistory[tracking.cpuUsageHistory.length - 1]?.value || 0;
+        const latestGpu = tracking.gpuUsageHistory[tracking.gpuUsageHistory.length - 1]?.value || 0;
+        const latestNetwork = tracking.networkUsageHistory[tracking.networkUsageHistory.length - 1]?.value || 0;
+        
+        // Estimate power consumption (simplified model)
+        consumption.cpu = latestCpu * 500; // mAh/hour
+        consumption.gpu = latestGpu * 800; // mAh/hour
+        consumption.screen = 300; // Base screen consumption
+        consumption.network = latestNetwork * 200; // mAh/hour
+        
+        consumption.total = consumption.baselinePower + 
+                           consumption.cpu + 
+                           consumption.gpu + 
+                           consumption.screen + 
+                           consumption.network;
+    }
+    
+    /**
+     * Update battery usage statistics
+     */
+    updateBatteryUsage() {
+        const usage = this.batteryMonitoring.usage;
+        const consumption = this.batteryMonitoring.powerConsumption;
+        
+        // Update current drain
+        usage.currentDrain = consumption.total;
+        
+        // Add to history
+        usage.drainHistory.push({
+            timestamp: Date.now(),
+            drain: usage.currentDrain
+        });
+        
+        // Limit history size
+        if (usage.drainHistory.length > usage.maxHistorySize) {
+            usage.drainHistory.shift();
+        }
+        
+        // Calculate average drain
+        if (usage.drainHistory.length > 0) {
+            const totalDrain = usage.drainHistory.reduce((sum, entry) => sum + entry.drain, 0);
+            usage.averageDrain = totalDrain / usage.drainHistory.length;
+        }
+        
+        // Estimate remaining time
+        const battery = this.batteryMonitoring.battery;
+        if (usage.averageDrain > 0 && !battery.charging) {
+            const remainingCapacity = battery.level * 3000; // Assume 3000mAh battery
+            usage.estimatedRemaining = remainingCapacity / usage.averageDrain;
+        }
+    }
+    
+    /**
+     * Evaluate battery health
+     */
+    evaluateBatteryHealth() {
+        const health = this.batteryMonitoring.health;
+        
+        // Simple health estimation based on usage patterns
+        // In reality, this would use more sophisticated algorithms
+        
+        if (health.cycleCount > 500) {
+            health.capacity = Math.max(70, 100 - (health.cycleCount - 500) * 0.1);
+        }
+        
+        // Estimate temperature based on power consumption
+        const consumption = this.batteryMonitoring.powerConsumption;
+        health.temperature = 25 + (consumption.total / 1000) * 15; // Simple model
+    }
+    
+    /**
+     * Optimize power consumption
+     */
+    optimizePowerConsumption() {
+        const consumption = this.batteryMonitoring.powerConsumption;
+        const threshold = 1000; // mAh/hour threshold
+        
+        if (consumption.total > threshold) {
+            console.log('[MobileBatteryOptimizer] High power consumption detected, applying optimizations');
+            this.applyAggressiveOptimizations();
+        }
+    }
+    
+    /**
+     * Apply basic battery optimizations
+     */
+    applyBasicOptimizations() {
+        const opts = this.batteryConfig.optimizations;
+        
+        opts.backgroundThrottling = true;
+        opts.networkOptimization = true;
+        opts.cacheAggressive = true;
+        
+        console.log('[MobileBatteryOptimizer] Basic optimizations applied');
+    }
+    
+    /**
+     * Apply aggressive battery optimizations
+     */
+    applyAggressiveOptimizations() {
+        const opts = this.batteryConfig.optimizations;
+        
+        opts.reducedFrameRate = true;
+        opts.cpuThrottling = true;
+        opts.suspendInactive = true;
+        opts.minimizeWakeups = true;
+        
+        console.log('[MobileBatteryOptimizer] Aggressive optimizations applied');
+    }
+    
+    /**
+     * Enable all battery optimizations
+     */
+    enableAllOptimizations() {
+        const opts = this.batteryConfig.optimizations;
+        
+        Object.keys(opts).forEach(key => {
+            opts[key] = true;
+        });
+        
+        console.log('[MobileBatteryOptimizer] All optimizations enabled');
+    }
+    
+    /**
+     * Notify application of battery state
+     */
+    notifyBatteryState(state) {
+        // This would typically dispatch events or call callbacks
+        console.log(`[MobileBatteryOptimizer] Battery state notification: ${state}`);
+        
+        // Dispatch custom event
+        if (typeof window !== 'undefined') {
+            const event = new CustomEvent('batteryStateChange', {
+                detail: { state, level: this.batteryMonitoring.battery.level }
+            });
+            window.dispatchEvent(event);
+        }
+    }
+    
+    /**
+     * Get battery optimization statistics
+     */
+    getBatteryStatistics() {
+        return {
+            battery: this.batteryMonitoring.battery,
+            usage: this.batteryMonitoring.usage,
+            powerConsumption: this.batteryMonitoring.powerConsumption,
+            health: this.batteryMonitoring.health,
+            currentMode: this.batteryConfig.currentMode,
+            optimizations: this.batteryConfig.optimizations
+        };
+    }
+    
+    /**
+     * Set battery optimization callbacks
+     */
+    setBatteryCallbacks(callbacks) {
+        this.batteryCallbacks = callbacks;
+    }
+    
+    /**
+     * Get power mode settings
+     */
+    getPowerModeSettings(mode) {
+        return this.batteryConfig.powerModes[mode] || null;
+    }
+    
+    /**
+     * Dispose battery optimizer
+     */
+    dispose() {
+        try {
+            // Clean up monitoring intervals
+            // Clear history arrays
+            this.performanceTracking.frameRateHistory = [];
+            this.performanceTracking.cpuUsageHistory = [];
+            this.performanceTracking.gpuUsageHistory = [];
+            this.performanceTracking.networkUsageHistory = [];
+            
+            this.batteryMonitoring.usage.drainHistory = [];
+            
+            console.log('[MobileBatteryOptimizer] Battery optimizer disposed');
+        } catch (error) {
+            this.errorHandler.handleError(error, 'MobileBatteryOptimizer.dispose');
+        }
+    }
+}
